Harden parser types, loops, and predicates
- Widen Marker/Event positions to usize, removing casts
- Widen n_raw_tokens to u32, removing truncation
- Add checkpoint-based progress guards to 6 loops
- Centralize type keyword and direction predicates
- Document fuel pattern and Ident heuristic

+++ crates/lyra-parser/src/event.rs
@@ -5,11 +5,11 @@ use lyra_lexer::SyntaxKind;
 pub(crate) enum Event {
     Start {
         kind: SyntaxKind,
-        forward_parent: Option<u32>,
+        forward_parent: Option<usize>,
     },
     Finish,
     // Consume `n_raw_tokens` from the token stream (trivia + one significant).
     Token {
-        n_raw_tokens: u16,
+        n_raw_tokens: u32,
     },
 }
